Scala 3 port of LogHelper macros
